+++ frontend/src/apps/client/features/order-history/types.ts
@@ -22,10 +22,11 @@ export type OrderHistorySearchParams = {
 };
 
 /**
- * 날짜 범위 프리셋 콤보 값. `SelectInput`은 value가 빈 문자열인 옵션을 무효 처리해 제외하므로
+ * 날짜 범위 프리셋 콤보 값. `useDateRangePresetDraft`(공용 훅)의 타입을 그대로 재사용한다.
+ * `SelectInput`은 value가 빈 문자열인 옵션을 무효 처리해 제외하므로
  * "직접 선택"도 비어 있지 않은 sentinel 값(`direct`)을 사용한다.
  */
-export type OrderHistoryDatePresetKey = 'direct' | 'week' | 'month' | 'year';
+export type { DateRangePresetKey as OrderHistoryDatePresetKey } from '@/shared/hooks/useDateRangePresetDraft';
 
 /** 주문 상태 필터 콤보 값. `ALL`은 전체(필터 미적용)를 의미한다. */
 export type OrderHistoryStatusFilterKey = 'ALL' | OrderHistoryStatus;
